+++ lib/groupdate/scopes.rb
@@ -10,5 +10,15 @@ module Groupdate
         Groupdate::Magic.new(field, options).relation(args[0], self)
       end
     end
+
+    def group_by_period(period, field, options = {})
+      # to_sym is unsafe on user input, so convert to strings
+      permitted_periods = ((options[:permit] || Groupdate::FIELDS).map(&:to_sym) & Groupdate::FIELDS).map(&:to_s)
+      if permitted_periods.include?(period.to_s)
+        send("group_by_#{period}", field, options)
+      else
+        raise ArgumentError, "Unpermitted period"
+      end
+    end
   end
 end